+++ galois/ir/ir.hpp
@@ -3,4 +3,4 @@
 #include "galois/ir/constant.hpp"
 #include "galois/ir/io.hpp"
 #include "galois/ir/tensor.hpp"
-#include "galois/ir/view.hpp"
+#include "galois/ir/view/view.hpp"
