+++ galois/ir/view/view.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "galois/ir/view/bit_cast.hpp"
+#include "galois/ir/view/broadcast.hpp"
+#include "galois/ir/view/flatten.hpp"
+#include "galois/ir/view/slice.hpp"
+#include "galois/ir/view/squeeze.hpp"
+#include "galois/ir/view/squeeze_dim.hpp"
+#include "galois/ir/view/transpose.hpp"
+#include "galois/ir/view/unsqueeze_dim.hpp"
+#include "galois/ir/view/viewer.hpp"
