+++ galois/ir/view/viewer.hpp
@@ -1,0 +1,49 @@
+#pragma once
+
+#include "galois/ir/tensor.hpp"
+
+namespace galois::ir::view {
+
+class Viewer : public Instruction {
+   public:
+    static std::shared_ptr<Viewer> Create(std::shared_ptr<Tensor> ir_tensor,
+                                          Eigen::MatrixXi64 transform_matrix,
+                                          Eigen::VectorXi64 shift_vector) {
+        std::shared_ptr<Viewer> self(new Viewer);
+        self->ir_tensor = ir_tensor;
+        self->transform_matrix = transform_matrix;
+        self->type = ir_tensor->type;
+        self->shift_vector = shift_vector;
+        self->tag = "Viewer";
+        return self;
+    }
+
+    static std::shared_ptr<Viewer> Shift(std::shared_ptr<Tensor> ir_tensor,
+                                         Eigen::VectorXi64 shift_vector) {
+        auto tensor_rank = ir_tensor->type->shape.size();
+        auto identity_matrix = Eigen::MatrixXi64::Identity(tensor_rank, tensor_rank);
+        GALOIS_ASSERT(shift_vector.size() == tensor_rank);
+        return Create(ir_tensor, identity_matrix, shift_vector);
+    }
+
+    static std::shared_ptr<Viewer> Stride(std::shared_ptr<Tensor> ir_tensor,
+                                          Eigen::VectorXi64 stride_vector) {
+        auto tensor_rank = ir_tensor->type->shape.size();
+        GALOIS_ASSERT(stride_vector.size() == tensor_rank);
+        Eigen::MatrixXi64 transform_matrix = Eigen::MatrixXi64::Zero(tensor_rank, tensor_rank);
+        for (int64_t i = 0; i < tensor_rank; ++i) {
+            transform_matrix(i, i) = stride_vector[i];
+        }
+        return Create(ir_tensor, transform_matrix, Eigen::VectorXi64::Zero(tensor_rank));
+    }
+
+    void ApplyVisitor(std::shared_ptr<Visitor> interpreter) override {
+        interpreter->Visit(Cast<Viewer>(this->shared_from_this()));
+    }
+
+    Eigen::MatrixXi64 transform_matrix;
+    Eigen::VectorXi64 shift_vector;
+    std::shared_ptr<Tensor> ir_tensor = nullptr;
+};
+
+}  // namespace galois::ir::view
